requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.16.0
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.12.0
supabase==2.3.0
